refactor: split metrics collection into TCP and UDP
